requirements.txt
amqp==5.3.1
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
audioread==3.0.1
av==15.0.0
billiard==4.2.1
black==24.4.2
boto3==1.40.10
botocore==1.40.10
celery==5.3.6
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.10.3
ctranslate2==4.6.0
decorator==5.2.1
Django==5.0.3
django-cors-headers==4.3.1
django-ratelimit==4.1.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
factory-boy==3.3.0
Faker==37.5.3
faster-whisper==1.2.0
ffmpeg-python==0.2.0
filelock==3.18.0
flake8==7.0.0
flatbuffers==25.2.10
fsspec==2025.7.0
future==1.0.0
h11==0.16.0
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
MarkupSafe==3.0.2
mccabe==0.7.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9
numba==0.61.2
numpy==1.26.4
onnxruntime==1.22.1
openai-whisper==20231117
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
prompt_toolkit==3.0.51
protobuf==6.32.0
psycopg2==2.9.10
pycodestyle==2.11.1
pycparser==2.22
pydub==0.25.1
pyflakes==3.2.0
PyJWT==2.8.0
pyreadline3==3.5.4
pytest==8.2.0
pytest-cov==4.1.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.2
rest-framework-simplejwt==0.0.2
rpds-py==0.27.0
ruff==0.12.9
s3transfer==0.13.1
safetensors==0.6.2
scikit-learn==1.4.1.post1
scipy==1.16.1
sentence-transformers==5.1.0
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
sse-starlette==1.6.1
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.2.2
torchaudio==2.2.2
tqdm==4.67.1
transformers==4.55.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
webrtcvad==2.0.10
win32_setctime==1.2.0
